update vscode config; add macos install script and zshrc variants

+++ install/desktop/app-vscode.sh
@@ -1,15 +1,21 @@
-cd /tmp
-wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor >packages.microsoft.gpg
-sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg
-echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | sudo tee /etc/apt/sources.list.d/vscode.list >/dev/null
-rm -f packages.microsoft.gpg
-cd -
+if [[ $(uname) == 'Darwin' ]]; then
+  brew install --cask visual-studio-code
 
-sudo apt update -y
-sudo apt install -y code
+  mkdir -p ~/Library/Application\ Support/Code/User
+  cp ~/Code/mgyarmathy/setup/dotfiles/vscode/.config/Code/User/settings.json ~/Library/Application\ Support/Code/User/settings.json
+elif [[ $(uname) == 'Linux' ]]; then
+  cd /tmp
+  wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor >packages.microsoft.gpg
+  sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg
+  echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | sudo tee /etc/apt/sources.list.d/vscode.list >/dev/null
+  rm -f packages.microsoft.gpg
+  cd -
 
-mkdir -p ~/.config/Code/User
-cp ~/.local/share/omakub/configs/vscode.json ~/.config/Code/User/settings.json
+  sudo apt update -y
+  sudo apt install -y code
 
-# Install default supported themes
-code --install-extension enkia.tokyo-night
+  mkdir -p ~/.config/Code/User
+  cp ~/Code/mgyarmathy/setup/dotfiles/vscode/.config/Code/User/settings.json ~/.config/Code/User/settings.json
+fi
+
+code --install-extension ms-vscode.Theme-TomorrowKit
